requirements.txt
alembic==1.13.3
Flask==3.0.3
Flask_Cors==5.0.0
Flask_HTTPAuth==4.8.0
Flask_Migrate==4.0.7
Flask_RESTful==0.3.10
flask_sqlalchemy==3.1.1
marshmallow==3.23.1
marshmallow_enum==1.5.1
password_strength==0.0.3.post2
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
SQLAlchemy==2.0.36
Werkzeug==3.1.2

requests~=2.32.3